Added application

+++ app/Http/Controllers/API/ProgrammeController.php
@@ -5,8 +5,12 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\ProgrammeResource;
 use App\Models\Programme;
+use App\Models\Sala;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Http\Request;
+use App\Models\User;
+
 
 class ProgrammeController extends Controller
 {
@@ -18,7 +22,7 @@ class ProgrammeController extends Controller
     public function index()
     {
         $clase=Programme::all();
-        return response(['programmes' => ProgrammeResource::collection($clase),'message'=>'Incarcate cu succes'],200);
+        return response(['programmes' => ($clase),'message'=>'Incarcate cu succes'],200);
     }
 
     /**
@@ -29,22 +33,55 @@ class ProgrammeController extends Controller
      */
     public function store(Request $request)
     {
-        $data=$request->all();
+        $user=User::where('id',auth()->user()->getAuthIdentifier())->first();
+        if($user->adminToken!=null) {
+            $data = [$request->class, $request->sala_id, $request->locuri, $request->from, $request->to];
+            $data = $request->all();
+            $validator = Validator::make($data, [
+                'class' => 'required|max:50|string',
+                'sala_id' => 'required|integer',
+                'locuri' => 'required|integer',
+                'from' => 'required',
+                'to' => 'required'
+            ]);
 
-        $validator=Validator::make($data,[
-            'clasa' => 'required|max:50',
-            'sala' =>'required',
-            'locuri' =>'required',
-            'from' =>'required',
-            'to'=>'required'
-        ]);
+            $salavalidator = Sala::where('id', $request->sala_id)->first();
+            if ($salavalidator == null) {
+                return response(['message' => 'sala invalida'], 400);
+            }
 
-        if($validator->fails()){
-            return response(['error'=>$validator->errors(),'Validation Error']);
+            if ($validator->fails()) {
+                return response(['error' => $validator->errors(), 'Validation Error'], 403);
+            }
+
+            //
+            $fromvalidation = DB::table('programmes')
+                ->select()
+                ->where('sala_id', $request->sala_id)
+                ->whereBetween('from', [$request->from, $request->to])
+                ->first();
+
+            $tovalidation = DB::table('programmes')
+                ->select()
+                ->where('sala_id', $request->sala_id)
+                ->whereBetween('to', [$request->from, $request->to])
+                ->first();
+
+            if ($fromvalidation == null && $tovalidation == null) {
+
+                $programme = new Programme();
+                $programme->class = $request->class;
+                $programme->locuri = $request->locuri;
+                $programme->sala_id = $request->sala_id;
+                $programme->from = $request->from;
+                $programme->to = $request->to;
+                $programme->save();
+                return response(['programme' => new ProgrammeResource($programme), 'message' => 'Creat cu succes'], 200);
+
+            }
+            return response(['message' => 'Sala este ocupata in intervalul orar'], 400);
         }
-
-        $programme=Programme::create($data);
-        return response(['programme'=>new ProgrammeResource($programme),'message'=>'Creat cu succes'],200);
+        else return response(['message','forbidden'],403);
     }
 
     /**
@@ -53,9 +90,22 @@ class ProgrammeController extends Controller
      * @param  \App\Models\Programme  $programme
      * @return \Illuminate\Http\Response
      */
-    public function show(Programme $programme)
+    public function show(Request $request)
     {
-        return response(['programme'=>new ProgrammeResource($programme), 'message'=>'Afisat cu succes'],200);
+        $data = $request->all();
+        $validator = Validator::make($data, [
+            'programme_id' => 'required',
+        ]);
+        if($validator->fails()){
+            return response(['message'=>'There has been an error'],400);
+        }
+
+        $programme=Programme::where('id',$request->programme_id)->first();
+        if($programme==null){
+            return response(['message'=>'The requested programme was not found'],404);
+        }
+
+        return response(['programme'=>$programme, 'message'=>'Afisat cu succes'],200);
     }
 
     /**
@@ -78,10 +128,14 @@ class ProgrammeController extends Controller
      * @param  \App\Models\Programme  $programme
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Programme $programme)
+    public function destroy(Request $request)
     {
-        $programme->delete();
-
-        return response(['message'=>"Sters"]);
+        $user=User::where('id',auth()->user()->id)->first();
+        if($user->adminToken!=null) {
+            $programme=Programme::where('id',$request->id)->first();
+            $programme->delete();
+            return response(['message' => "Sters"]);
+        }
+        else return response(['message'=>'Forbidden'],403);
     }
 }
